pyproject.toml
[tool.poetry]
name = "bboard"
version = "0.1.0"
description = ""
authors = ["jshuckbot <gorka_ivan@mail.ru>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
django = "4.2"
django-bootstrap5 = "^23.4"
easy-thumbnails = "^2.8.5"
django-cleanup = "^8.1.0"
pillow = "^10.3.0"
django-simple-captcha = "^0.6.0"
djangorestframework = "^3.15.1"
django-cors-headers = "^4.3.1"


[tool.poetry.group.dev.dependencies]
isort = "^5.13.2"
black = "^24.3.0"
autoflake = "^2.3.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
